python-be/requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==3.1.3
flask-restx==1.3.0
python-dotenv==1.0.1
gunicorn==21.2.0
requests
sqlalchemy
flask_sqlalchemy
pymysql
flask_cors
python-jose[cryptography]
azure-cognitiveservices-speech
babel